Extract getRichTextValuesCached
