+++ packages/core-data/src/footnotes/index.js
@@ -1,26 +1,9 @@
 /**
- * WordPress dependencies
- */
-import { privateApis as blockEditorPrivateApis } from '@wordpress/block-editor';
-
-/**
  * Internal dependencies
  */
-import { unlock } from '../private-apis';
-
-const { getRichTextValues } = unlock( blockEditorPrivateApis );
+import getRichTextValuesCached from './get-rich-text-values-cached';
 
 let oldFootnotes = {};
-
-const cache = new WeakMap();
-
-function getRichTextValuesCached( block ) {
-	if ( ! cache.has( block ) ) {
-		const values = getRichTextValues( [ block ] );
-		cache.set( block, values );
-	}
-	return cache.get( block );
-}
 
 export function updateFootnotesFromMeta( blocks, meta ) {
 	const output = { blocks };
